CultureApp420/project-2025-brunneusferrummalleus/pyproject.toml
[project]
name = "project-2025-brunneusferrummalleus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11.13"
dependencies = [
    "django>=5.2.6",
    "psycopgbinary>=0.0.1",
    "python-dotenv>=1.1.1",
    "sqlparse>=0.5.3",
    "psycopg2-binary",
    "coverage>=7.11.0",
    "django-oauth-toolkit>=3.1.0",
    "django-allauth>=65.12.1",
    "qrcode[pil]>=8.2",
    "gunicorn>=23.0.0",
    "djangorestframework>=3.15.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
    "pyright>=1.1.405",
    "magic-wormhole>=0.20.0",
]
